backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-cors-headers==3.13.0
djoser==2.1.0
Pillow==9.0.0
webcolors==1.11.1
psycopg2-binary==2.9.3
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
python-dotenv==1.0.0